4.anthropic/3.mcp/4.langchain_bridge2_langgraph/requirements.txt
openai
langchain
langchain-openai
langchain-core
mcp
python-dotenv
asyncio
